update on card component and category component and prodect card component;
