+++ src/components/Card/Types.ts
@@ -1,0 +1,8 @@
+export interface Card{
+    img:string;
+    name:string;
+    salary?:number;
+    valueRating?:number;
+    discountValue?:number;
+    
+}
