--- src/components/CategroiesCard/CategroiesCard.styled.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-export const ContinerCategroiesCard = styled('div')(({ theme }) => ({
-    display: 'flex',
-    width: '70%',
-    gap:'1%',
-    
-  
-    }));
-  
-
