--- src/components/ProdectsCard/ProdectsCard.styled.ts
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-
-export const ContinerCategroiesCard = styled('div')(({ theme }) => ({
-    display: 'flex',
-    width: '70%',
-    gap:'3%',
-    
-  
-    }));
-    export const SittingContainer = styled.div`
-width:70%;
-display:flex;
-align-items:center;
-justify-content:space-between;
-margin-left:10%;`;
-  
-
